+++ src/afterfact.rs
@@ -1500,6 +1500,7 @@ mod tests {
                     quiet_errors: false,
                     config: Path::new("./rules/config").to_path_buf(),
                     verbose: false,
+                    json_input: false,
                 },
                 profile: None,
                 output: Some(Path::new("./test_emit_csv.csv").to_path_buf()),
@@ -1523,7 +1524,6 @@ mod tests {
                 html_report: None,
                 no_summary: true,
             },
-            json_input: false,
             geo_ip: None,
         });
         let dummy_config = Some(Config {
@@ -1567,6 +1567,7 @@ mod tests {
                     quiet_errors: false,
                     config: Path::new("./rules/config").to_path_buf(),
                     verbose: false,
+                    json_input: false,
                 },
                 profile: None,
                 output: Some(Path::new("./test_emit_csv.csv").to_path_buf()),
@@ -1821,6 +1822,7 @@ mod tests {
                 quiet_errors: false,
                 config: Path::new("./rules/config").to_path_buf(),
                 verbose: false,
+                json_input: false,
             },
             profile: None,
             output: Some(Path::new("./test_emit_csv.csv").to_path_buf()),